move config file code from -stats and token into a separate file
this is so it can be reused by other apps

see https://github.com/Strubbl/wallabag-stats/issues/1
